requirements.txt
absl-py==0.9.0
astor==0.8.1
cloudpickle==1.1.1
cycler==0.10.0
decorator==4.4.1
gast==0.2.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
imageio==2.6.1
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.1.2
matplotlib2tikz==0.7.6
numpy==1.18.0
opt-einsum==3.1.0
Pillow==7.0.0
protobuf==3.11.2
pyparsing==2.4.6
python-dateutil==2.8.1
remi-uob-style==2019.2
scikit-learn==0.22.1
scipy==1.4.1
six==1.13.0
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
tensorflow-probability==0.8.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
xlrd==1.2.0
